+++ src/main/java/com/hailtosg/reactive/controller/v1/ItemController.java
@@ -21,12 +21,6 @@ public class ItemController {
         itemReactiveRepository = itemRepository;
     }
 
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<String> handleRte (RuntimeException exception) {
-        log.error("RTE :", exception);
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(exception.getMessage());
-    }
-
     @GetMapping(ITEMS_END_POINT_V1)
     public Flux<Item> getAllItems(){
         return itemReactiveRepository.findAll();
@@ -42,7 +36,7 @@ public class ItemController {
     @GetMapping(ITEMS_END_POINT_V1 + "/rte")
     public Flux<Item> getRtE(){
         return itemReactiveRepository.findAll()
-                .concatWith(Mono.error(new RuntimeException("RTE!!!")));
+                .concatWith(Mono.error(new RuntimeException("RTE occurred")));
     }
 
     @PostMapping(ITEMS_END_POINT_V1)
